Update patch set 1
Patch Set 1: Commit-Queue+1

Piotr: I'm sure that you know more about Bazel than I do. (I've not had a chance to test this version of the BUILD file on macOS either, although I will tomorrow.)

Sadly it appears that CPU strings are as much of a mess in Bazel as they are externally.

Patch-set: 1
Label: Commit-Queue=+1