Update patch set 2
Patch Set 2:

Regarding WORKSPACE, yeah, it's fine to add it, it works fine with external repositories.

Also, you might want to add this line to it:

    workspace(name = "boringssl")

This will make sure that it's always going to be included as "boringssl" external repository and not under different name (for the sake of consistency, so that external BUILD files would always depend on it as "@boringssl//:crypto").

This is obviously optional, but it seems that Bazel (name=io_bazel), Brotli & TensorFlow (name=org_tensorflow) use this.

Patch-set: 2