BUILD: add correct defines for shared-library building.
If we're going to set -fvisibility=hidden, we should also set the
defines needed to include the visibility annotations.

Since Bazel doesn't have a shared vs static concept that's exposed to
the build rules (as far as I know), this means that the annotations will
be set even for a static build. But that appears to be fine, at least on
Linux (and Bazel doesn't build on Windows so we don't have to worry
about that).

Change-Id: Ic7d64a553da48cfb9cf5460d26254de7e105fd65