Update patch set 3
Patch Set 3:

(1 comment)

I think that this is ultimately the right direction, but I don't think that we can land this change for now. Rather it'll take a bit of work for someone at Google to roll this out.

We've not had the situation before where the PRNG will block if the entropy pool hasn't initialised yet. Certainly we hope that nothing is running with an immature pool, but we haven't tested it so we can't be sure of that. There is a possibility that, somewhere, some system is running processes on a freshly started machine with an immature pool and this will cause its startup to block, which might well prevent enough happening for the pool to fill and thus deadlocking. That could be a real mess.

We would need to test and communicate this change extensively, which I'm afraid will take some time.

Patch-set: 3
Label: Code-Review=0