Update patch set 2
Patch Set 2:

In that we should probably move the logic into ssl_rsa in a separate method, instead of having the handshake machinery need to call ssl3_cert_verify_hash and then digest/sign it. That way we keep all the signing related things in that file instead of split.

Patch-set: 2
Label: Code-Review=0