Update patch set 2
Patch Set 2:

> In that we should probably move the logic into ssl_rsa in a
 > separate method, instead of having the handshake machinery need to
 > call ssl3_cert_verify_hash and then digest/sign it. That way we
 > keep all the signing related things in that file instead of split.

What would the method take in as input? It can't be the handshake transcript since it involves the master secret. We also need the hash to be separated out because public and private key halves deal with it separately. Not sure I see what this buys us here.

This is really just about us computing the hash a third time. ssl3_cert_verify_hash knows it (it internally dispatches on sigalg), the handshake knows it (it conditions on pkey->type), and yet the handshake still needs to call tls12_get_hash again, which is silly.

Turns out the call isn't needed, so I figured we could remove it. Or, if you prefer, cert_verify_hash can give back an output parameter for what hash it used.

Patch-set: 2
Label: Code-Review=0