Update patch set 1
Patch Set 1:

> Just to close this request with a comment.. Gong and I talked about
 > this.. borringssl is used in grpc, which is used in hyperledger
 > (blockchain project). For the moment, we created our own fork of at
 > https://github.com/linux-on-ibm-z/boringssl to fix the build-break
 > for hyperledger. That's the location where we already maintain (and
 > upstream) other deserving projects (i.e. golang, which I am
 > contributing P256 too).

If gRPC make a request of us to support s390x then that could be a viable route for you since they are a Google-based consumer. However, they would then have to provide us with buildbot support and some answer to the problem of how to debug any failures without hardware.

Patch-set: 1