Update patch set 4
Patch Set 4:

I think that, in order to validate this internally, we would need to do something like the following:

In the once that initialises the PRNG (with some small, fixed probability) call getrandom with GRND_NONBLOCK and log a message if it fails. (I.e. if the entropy pool isn't ready yet.)

We can monitor for certain kinds of log messages across our fleet and watch to see whether this is going to be an issue. Over weeks, we'll ramp up the probability until we have good coverage.

But I'm afraid that the practical upshot is that we'll probably have to manage that ourselves I'm afraid, before this can be landed and enabled by default.

Might I ask what your motivation is for this change? Do you have a specific need in mind?

Patch-set: 4
Label: Code-Review=0