Add a ssl_hs_write_message_and_flush wait mode.
Eventually write_message will be gone and finish_message will simply
take care of that, which will let us cut down on all states within a
flight (though it may turn out cleaner not to) and we would be able to
switch cipher states as soon as finish_message returns. But, in the
meantime, add the combined wait mode to make things slightly simpler.

Change-Id: Iebaa388ccbe7fcad48c1b2256e1c0d3a7c9c8a2a