Make assembly clang-as compatible
Most of the __APPLE__ labels are actually looking for clang. Switch them
to express what is actually meant.

Additionally the SHA-256 assembly files apparently only are assembled in
thumb mode before, but there is a use of "ADRL" pseudo-instruction that
can be replaced with a .word calculation of the label offset as
laid out in a bug report here:

https://llvm.org/bugs/show_bug.cgi?id=24350#c22

Change-Id: I78e684fd737eee6a55a303695392ed2684c7c25b