Update patch set 3
Patch Set 3:

One thing to keep in mind is that BN_mod_inverse is a variable-time algorithm so it will have (sometimes dramatically) different timings for different inputs.

Also, I seem to remember that on Windows BN_mod_inverse is much slower than on Linux. I'll look to see if that is the case. I think it might be slow because it is using BN_div_word (or whatever it's called) instead of the single `idiv` instruction for division, and maybe because in general a bunch of asm BN optimizations are GCC-specific?

Patch-set: 3
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0