Reject stray post-Finished messages in DTLS.
This is in preparation for switching finish_handshake to a
release_current_message hook. finish_handshake in DTLS is also
responsible for releasing any memory associated with extra messages in
the handshake.

Except that's not right and we need to make it an error anyway. Given
that the rest of the DTLS dispatch layer already strongly assumes there
is only one message in epoch one, putting the check in the fragment
processing works fine enough. Add tests for this.

This will certainly need revising when DTLS 1.3 happens (perhaps just a
version check, perhaps bringing finish_handshake back as a function that
can fail... which means we need a state just before SSL_ST_OK), but I'm
unclear on how DTLS post-handshake messages work, so let's do the easy
thing for now and add a test.

This removes the sequence number reset in the DTLS code. That reset
never did anything. It would if we ever implemented renego, but we
won't. Hopefully DTLS 1.3 will not include that reset (it was a
mistake).

BUG=83

Change-Id: I33d645a8550f73e74606030b9815fdac0c9fb682