Add a shim for ssl/test/runner using the bogo implementation.
The idea here is to test the Go implementation against itself (as if it
were the C implementation). To pretend to be the C implementation for
purposes of testing, the shim has a map of bogo error strings to error
strings that the runner is matching against. Most of the flags for the
shim are currently unimplemented.

Out of the 2461 tests that currently exist, 2016 tests are skipped (due
to unimplemented flags on the shim), and another 109 fail, leaving 336
that are passing.  Some of the failures appear to be legitimate
failures, while others could be not correctly mapping bogo error
strings. (And there might be some errors getting masked from maping one
bogo error to multiple errors.)

Change-Id: Id72baf4f7ae899be16bde00e4f0ccc4b1cc997e8