Update patch set 15
Patch Set 15:

When I wrote this patch, I did so a a purely mechanical transformation: All I did was merge the logic in the `#if` branch with the similar logic in the `#else` branch, without otherwise trying to improve anything.

For a while, I wondered why there was hesitancy to accept this patch. But, recently, I started working on improving the actual logic beyond just removing `#if`...`#else`. Now I realize that if you actually are trying to understand what the code is doing, beyond just making sure the new code does exactly what the old code does, you're going to be completely lost, unless you understand the shortcuts that the code is using, which aren't documented in the code.

For example, it would be instructive to know that the subtraction by one is actually inverting the sign of the number in two's-complement notation. I did not know this until yesterday.

Now I am working on completely replacing this logic with better logic and also documenting the shortcut. I'm not sure whether it makes sense to land this patch as an intermediate step or not.

Patch-set: 15
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0