Update patch set 15
Abandoned

No worries; I completely understand. Anyway, the code was hard to figure out even once I understood the math. My other changeset hopefully makes things much clearer, and it is a replacement of this one (it doesn't depend on this one), so let's drop this one.

Patch-set: 15
Status: abandoned