Update patch set 15
Patch Set 15:

> I noticed that you removed the comment I added about why we use
 > BN_BITS2 instead of LG_LITTLE_R when calculating RR (mod n). I
 > still am not sure why BN_BITS2 works there and LG_LITTLE_R doesn't.
 > LG_LITTLE_R seems more clearly right to me, but the tests fail on
 > x86 if I use LG_LITTLE_R instead. If you know of an explanation as
 > to why, it would be great to get it.

Ah, I only removed the comment because it looked like it was a left-over. (LG_LITTLE_R wasn't defined in that scope.)

But LG_LITTLE_R will be 64 on x86 while BN_BITS2 will be 32, right? It's normal that R is the minimal number of limbs that will contain the modulus and I expect that some of the code assumes that there'll only be a single limb of overhead. Perhaps I'm missing something however.

Patch-set: 15