Update patch set 15
Patch Set 15:

> But LG_LITTLE_R will be 64 on x86 while BN_BITS2 will be 32, right?

Yes.

> It's normal that R is the minimal number of limbs that will contain the modulus and I expect that some of the code assumes that there'll only be a single limb of overhead. Perhaps I'm missing something however.

On x86 (at least), when using the SIMD code, limbs are effectively 64-bits. So, it would be interesting to study some moduli that result in an RR that is an odd number of 32-bit limbs (i.e. not evenly divisible into 64-bit limbs) and see if there is anything problematic. Potential problems linclude a one-limb buffer read overflow or similar in the ASM code reading RR, and/or a difference in behavior between the asm code for reduction and BN_from_montgomery_word. Regardless, I didn't change the semantics of this with this CL, so we can defer that investigation for later, and it is OK to remove that comment.

Patch-set: 15