Update patch set 1
Patch Set 1:

Hrm. If you're messing around with the legacy ASN.1 stack, mind just matching the indentation in OpenSSL 1.0.2's copy. This is closer to it, but still a little different.

Sorry, this is our bad for not documenting this anywhere. (I have a queue of docs I need to write.) But I wouldn't waste much time on these directories (crypto/{asn1,pem,x509,x509v3}) if I were you. This stuff is some of the lowest quality code in all of OpenSSL, and that's saying something. It's really beyond saving.

We've got a much much simpler and less crappy ASN.1 parser in crypto/bytestring that we've transitioning most internal bits over to using. What exactly is to be done with the X.509 bits is still a bit in the air. (On the Chromium side we're working on a new PKI library. That'll probably be involved in the final story.)

At the moment, we're mostly just leaving it in a corner and pulling changes from OpenSSL 1.0.2. So we actually want to just minimize divergence to keep applying patches simple-ish. (Though we don't really match either. We forked prior to their reformatting, ran their script at some point, and didn't pay a whole lot of attention to their manual fix-ups.)

Patch-set: 1