Factor out common logic in bn_mod_inverse_*.
This is in preparation for |bn_mod_inverse_odd| being exposed for use
outside of crypto/bn/gcd.c. The goal is to make |bn_mod_inverse_odd|'s
signature similar to |BN_mod_inverse_blinded|.

The logic for reducing the final result isn't factored out because
the yet-to-be-created |BN_mod_inverse_odd| will need to do it itself.

Change-Id: Iaecb79fb17d13c774c4fb6ade8742937780b0006