Update patch set 4
Patch Set 4:

Thinking about this patch some more:

Is it really viable long-term to do this differently than OpenSSL upstream? I don't think so. Imagine this coding pattern:

   int is_negative = BN_is_negative(x);
   
   <do lots of calculations that don't change
   the sign of x, but may change it to zero, using
   things other than the public BN_* API.>

   if (is_negative) {
     /* potential TOCTOU problem. */
   }

So, at a minimum it seems very important to get OpenSSL to take this change too, so that they don't add new code that uses the above pattern.

And, probably we need to audit that the above pattern doesn't occur. But, that seems hard to do.

Patch-set: 4