Update patch set 4
Patch Set 4:

To clarify, my concern is that there could be code that depends on `bn_correct_top` leaving the sign unchanged. i.e. calls to bn_correct_top in the "do lots of calculations that don't change
the sign of x, but may change it to zero..." part.

I did just go through all the uses of bn_correct_top in BoringSSL. It turns out there are not that many and they all look safe w.r.t. this change.

Patch-set: 4