Update patch set 8
Patch Set 8:

I'm not sure about the changes to the tests I made in the latest update. It seems like if this ever lands, there's no longer much point in testing BN_div specially with BN_FLG_CONSTTIME, but maybe you want to just because it's a regression test.

Patch-set: 8