Update patch set 8
Patch Set 8:

Thinking about this some more: This change primarily just avoids some  short-circuiting logic that actually should rarely be used. In the course of ECDH operations, BN_div shouldn't be called nearly enough to affect performance +/- 3% or more. If it were called frequently that would be a sign of a more serious problem. But, even if it were called frequently and that was expected, we would expect it to be very rare that the short-circuit case would be hit during an ECDH operation.

IF we measure EC_GROUP_new_whatever in speed.cc and compare it to, say, EC_POINT_dbl, we should get a sense as to whether optimizing EC_GROUP_new_whatever would make a significant difference. I would expect EC_POINT_dbl and EC_POINT_add, and in particular BN_mod_mul_montgomery, to completely dominate everything w.r.t. any performance profile.

Of course, that could all be wrong, but I think the next step should probably be to find a way to measure the performance with less noise.

Patch-set: 8
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0