Update patch set 2
Patch Set 2:

If one wanted deterministic signatures, why not specify a salt length of zero? Is there any security difference between a fixed salt of length > 0 and an empty salt?

Also, doing this globally worries me a little. Consider a protocol that replies to queries with a randomised encryption of a message and a PSS signature of that message. With this change, the signature would allow an attacker to build a mapping of signature value to reply contents and "decrypt" other replies. This is the reason why the (EC)DSA nonce generation still includes a random element even though it also hashes the secret and message.

Patch-set: 2