Update patch set 2
Patch Set 2:

> If one wanted deterministic signatures, why not specify a salt length of zero? Is there any security difference between a fixed salt of length > 0 and an empty salt?

Some implementations may require a non-empty salt since some protocols, like the current TLS 1.3 draft and some other protocol drafts at IETF, mandate a salt of a particular length.

I will need some time to think about the rest of your message. Are you assuming that the signing and encryption happens with the same RSA key? IIUC, you are basically saying that the security analysis of signature schemes that is typically done assumes that there is a signing oracle but doesn't assume that there are other oracles, in particular an encryption or decryption oracle.

Regardless, I do think it would be neat, and somewhat useful, to find a way to generate PSS signatures without relying on the PRNG. This was an idea I just came up with this weekend.

(Regarding the way the ECDSA nonces are randomized, I do have some concerns about the construct used but I also hadn't realized that there was a specific concern in mind beyond pure defense-in-depth thinking.)

Patch-set: 2