Update patch set 2
Patch Set 2:

Adam, now I think I understand what you mean, and that it isn't required that the RSA key be the same for the encryption and signing. Off the top of my head, I don't think it is a good idea for anybody to rely on the nondeterminism of signatures like that. That doesn't mean that people don't rely on it and it doesn't mean that there aren't any tough cases where one might feel compelled to do so. Without knowing of any real systems that make such assumptions of nondeterminism, and given the widespread existence of such systems seems unlikely to cross my path since almost nobody is using PSS anyway (AFAICT), I still think moving to a deterministic approach is better.

It is also true that I think it is better to just specify that there is no salt (sLen == 0) than to say there is a salt but then use a fixed one. I'd rather submit a path that just rejects all non-zero `sLen`. But, like I said, I suspect that there will be recipients of signatures that expect a salt of the digest output length and this is my attempt to interop with those implementations.

Definitely this is a questionable proposal. No hard feelings if you decide it isn't something to include in BoringSSL.

Patch-set: 2
Label: Code-Review=0